separated routes from the main file
