+++ example/lib/main.dart
@@ -8,37 +8,8 @@ import 'package:flutter/material.dart';
 import 'package:group_button/group_button.dart';
 
 import 'examples/button_builder_example/example.dart';
+import 'routes.dart';
 
-class Routes {
-  /// Use [GroupButtonExtendedExample]
-  /// to check all package [GroupingType]'s in one place
-  static const groupButtonExtendedExample = '/GroupButtonExtendedExample';
-
-  /// Use [StylesExample]
-  /// to check all package style's configuration
-  static const stylesExample = '/StylesExample';
-
-  /// Use [CustomizableExample]
-  /// to check package power
-  static const customizableExample = '/CustomizableExample';
-
-  /// Use [ButtonBuilderExample]
-  /// to check how button builder works
-  ///
-  /// In this example GroupButton work as
-  /// common flutter [Checkbox] and [Radio] buttons
-  static const buttonBuilderExample = '/ButtonBuilderExample';
-
-  /// Use [FullOptionsSelectedExample]
-  /// to check all [GroupButtonOptions]
-  static const fullOptionsSelectedExample = '/FullOptionsSelectedExample';
-
-  /// Use [GenericsExample]
-  /// to check how works generic button values
-  static const genericsExample = '/GenericsExample';
-
-  static const commonExample = '/CommonExample';
-}
 
 void main() {
   runApp(MyApp());
